+++ mogmogmanager/src/main/java/fr/skytech/application/services/IUserService.java
@@ -6,7 +6,7 @@ import fr.skytech.application.controller.dto.UserDTO;
 
 public interface IUserService {
 
-	public UserDTO findUserById(Integer id);
+	public UserDTO findUserById(Long id);
 	public UserDTO findUserByUsername(String username);
 	public List<UserDTO> findAll();
